Re-organised the root files in the folder structure
