+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
 import * as Sentry from '@sentry/browser';
-import ApolloContainer from './ApolloContainer';
+import App from './App';
 import * as serviceWorker from './serviceWorker';
 
 // initialising sentry for error logging
@@ -11,7 +11,7 @@ Sentry.init({
 });
 
 // eslint-disable-next-line react/jsx-filename-extension
-ReactDOM.render(<ApolloContainer />, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
